+++ forge/src/main/java/com/majruszsaccessories/Initializer.java
@@ -2,10 +2,16 @@ package com.majruszsaccessories;
 
 import com.majruszlibrary.item.CreativeModeTabHelper;
 import com.majruszsaccessories.items.CreativeModeTabs;
+import net.minecraft.client.renderer.texture.TextureAtlas;
 import net.minecraft.core.NonNullList;
+import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.client.event.TextureStitchEvent;
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.InterModComms;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.InterModEnqueueEvent;
@@ -32,6 +38,7 @@ public class Initializer {
 		MinecraftForge.EVENT_BUS.register( this );
 
 		FMLJavaModLoadingContext.get().getModEventBus().addListener( Initializer::onEnqueueIMC );
+		DistExecutor.unsafeRunWhenOn( Dist.CLIENT, ()->()->FMLJavaModLoadingContext.get().getModEventBus().addListener( Initializer::onTextureStitch ) );
 		CreativeModeTabHelper.createItemIconReplacer( CreativeModeTabs::getPrimaryIcons, CREATIVE_MODE_TAB.getDisplayName() );
 	}
 
@@ -46,4 +53,13 @@ public class Initializer {
 			.build()
 		);
 	}
+
+	@OnlyIn( Dist.CLIENT )
+	private static void onTextureStitch( TextureStitchEvent.Pre event ) {
+		final TextureAtlas map = event.getAtlas();
+		if( InventoryMenu.BLOCK_ATLAS.equals( map.location() ) ) {
+			event.addSprite( MajruszsAccessories.HELPER.getLocation( "slot/pocket" ) );
+		}
+	}
+
 }