+++ changelog.md
@@ -1,1 +1,2 @@
 - backported mod to 1.19.2
+- fixed missing pocket texture (reported by @SiverDX)